+++ apps/planners/domains/vanilla_team_sar.h
